use subclass
